requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
qrcode[pil]>=7.4
cryptography>=41.0.0
gunicorn>=21.2.0; sys_platform != 'win32'
requests>=2.31.0
Pillow>=10.0.0
pywebview>=5.0
pyinstaller>=6.0
